+++ pages/learn/get-started/[...slug].js
@@ -19,20 +19,14 @@
 import fs from "fs";
 import matter from "gray-matter";
 import React from "react";
-import ReactMarkdown from "react-markdown";
 import { Container, Col, Button, Offcanvas } from "react-bootstrap";
-import remarkGfm from "remark-gfm";
 import Image from "next-image-export-optimizer";
-import rehypeRaw from "rehype-raw";
 import Head from "next/head";
 import { Liquid } from "liquidjs";
 
-import { getHighlighter, setCDN } from "shiki";
-
-setCDN("https://unpkg.com/shiki/");
-
 import Layout from "../../../layouts/LayoutDocs";
 import LeftNav from "../../../components/common/left-nav/LeftNav";
+import MainContent from "../../../components/common/main-content/mainContent";
 import { prefix } from "../../../utils/prefix";
 import Toc from "../../../components/common/pg-toc/Toc";
 import LearnToc from "../../../utils/learn-lm.json";
@@ -134,80 +128,17 @@ export default function PostPage({
     return newContent;
   };
 
-  // Synatax highlighting
-  const HighlightSyntax = (code, language) => {
-    const [codeSnippet, setCodeSnippet] = React.useState([]);
-    if (language == "proto") {
-      language = "ballerina";
-    }
-    React.useEffect(() => {
-      async function fetchData() {
-        getHighlighter({
-          theme: "github-light",
-          langs: [
-            "bash",
-            "ballerina",
-            "toml",
-            "yaml",
-            "sh",
-            "json",
-            "graphql",
-            "sql",
-          ],
-        }).then((highlighter) => {
-          setCodeSnippet(highlighter.codeToHtml(code, language));
-        });
-      }
-      fetchData();
-    }, [code, language]);
-
-    return [codeSnippet];
-  };
-
   // Show mobile left nav
   const [show, setShow] = React.useState(false);
   const handleClose = () => setShow(false);
   const handleShow = () => setShow(true);
 
-  // Add id attributes to headings
-  const extractText = (value) => {
-    if (typeof value === "string") {
-      return value;
-    } else {
-      return value.props.children;
-    }
-  };
-
-  const scanArray = (array) => {
-    const newArray = array.map(extractText);
-    let newId = newArray
-      .join("")
-      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-      .toLowerCase();
-    newId = newId.replace(/ /g, "-");
-    return newId;
-  };
-
-  const getLink = (element, id) => {
-    if (element.tagName.toLowerCase() === "path")
-      element = element.parentElement;
-
-    const elementNodeList = document.querySelectorAll(`#${id}`);
-    const elementArray = Array.prototype.slice.call(elementNodeList);
-    const count = elementArray.indexOf(element.parentElement);
-
-    if (count === 0) {
-      location.hash = `#${id}`;
-    } else {
-      location.hash = `#${id}-${count}`;
-    }
-
-    navigator.clipboard.writeText(window.location.href);
-    element.parentElement.scrollIntoView();
-  };
-
   // Show page toc
   const [showToc, setShowToc] = React.useState(false);
+
+  const handleToc = (data) => {
+    setShowToc(data)
+  }
 
   return (
     <>
@@ -294,178 +225,10 @@ export default function PostPage({
 
             <p className="intro">{frontmatter.intro}</p>
 
-            <ReactMarkdown
-              components={{
-                h2({ node, inline, className, children, ...props }) {
-                  let id = "";
-                  setShowToc(true);
-                  if (children.length === 1) {
-                    id = children[0]
-                      .toString()
-                      .toLowerCase()
-                      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-                      .replace(/ /g, "-");
-                  } else {
-                    id = scanArray(children);
-                  }
-                  return (
-                    <h2 id={id} data-section={id} className="section">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="30"
-                        height="30"
-                        fill="currentColor"
-                        className="bi bi-link-45deg mdButton pe-2"
-                        viewBox="0 0 16 16"
-                        onClick={(e) => getLink(e.target, id)}
-                      >
-                        <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
-                        <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
-                      </svg>
-                      {children}
-                    </h2>
-                  );
-                },
-                h3({ node, inline, className, children, ...props }) {
-                  let id = "";
-                  setShowToc(true);
-                  if (children.length === 1) {
-                    id = children[0]
-                      .toString()
-                      .toLowerCase()
-                      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-                      .replace(/ /g, "-");
-                  } else {
-                    id = scanArray(children);
-                  }
-                  return (
-                    <h3 id={id} data-section={id} className="section">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="30"
-                        height="30"
-                        fill="currentColor"
-                        className="bi bi-link-45deg mdButton pe-2"
-                        viewBox="0 0 16 16"
-                        onClick={(e) => getLink(e.target, id)}
-                      >
-                        <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
-                        <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
-                      </svg>
-                      {children}
-                    </h3>
-                  );
-                },
-                h4({ node, inline, className, children, ...props }) {
-                  let id = "";
-                  setShowToc(true);
-                  if (children.length === 1) {
-                    id = children[0]
-                      .toString()
-                      .toLowerCase()
-                      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-                      .replace(/ /g, "-");
-                  } else {
-                    id = scanArray(children);
-                  }
-                  return (
-                    <h4 id={id} data-section={id} className="section">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="30"
-                        height="30"
-                        fill="currentColor"
-                        className="bi bi-link-45deg mdButton pe-2"
-                        viewBox="0 0 16 16"
-                        onClick={(e) => getLink(e.target, id)}
-                      >
-                        <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
-                        <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
-                      </svg>
-                      {children}
-                    </h4>
-                  );
-                },
-                h5({ node, inline, className, children, ...props }) {
-                  let id = "";
-                  setShowToc(true);
-                  if (children.length === 1) {
-                    id = children[0]
-                      .toString()
-                      .toLowerCase()
-                      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-                      .replace(/ /g, "-");
-                  } else {
-                    id = scanArray(children);
-                  }
-                  return (
-                    <h5 id={id} data-section={id} className="section">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="30"
-                        height="30"
-                        fill="currentColor"
-                        className="bi bi-link-45deg mdButton pe-2"
-                        viewBox="0 0 16 16"
-                        onClick={(e) => getLink(e.target, id)}
-                      >
-                        <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
-                        <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
-                      </svg>
-                      {children}
-                    </h5>
-                  );
-                },
-                h6({ node, inline, className, children, ...props }) {
-                  let id = "";
-                  setShowToc(true);
-                  if (children.length === 1) {
-                    id = children[0]
-                      .toString()
-                      .toLowerCase()
-                      .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
-                      .replace(/ /g, "-");
-                  } else {
-                    id = scanArray(children);
-                  }
-                  return (
-                    <h6 id={id} data-section={id} className="section">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="30"
-                        height="30"
-                        fill="currentColor"
-                        className="bi bi-link-45deg mdButton pe-2"
-                        viewBox="0 0 16 16"
-                        onClick={(e) => getLink(e.target, id)}
-                      >
-                        <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
-                        <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
-                      </svg>
-                      {children}
-                    </h6>
-                  );
-                },
-                code({ node, inline, className, children, ...props }) {
-                  const match = /language-(\w+)/.exec(className || '')
-                  return inline ?
-                    <code className={className} {...props}>
-                      {children}
-                    </code>
-                    : match ?
-                      <div dangerouslySetInnerHTML={{ __html: HighlightSyntax(String(children).replace(/\n$/, ''), match[1].toLowerCase()) }} />
-                      : <pre className='default'>
-                        <code className={className} {...props}>
-                          {children}
-                        </code>
-                      </pre>
-                }
-              }}
-              remarkPlugins={[remarkGfm]}
-              rehypePlugins={[rehypeRaw]}
-            >
-              {AddLiquid(content)}
-            </ReactMarkdown>
+            <MainContent
+              content={AddLiquid(content)}
+              handleToc={handleToc}/>
+
           </Container>
         </Col>
         <Col sm={2} xxl={3} className="pageToc d-none d-sm-block">
