+++ utils/rearrangeFiles.sh
@@ -3,6 +3,7 @@ cp -r ./out/learn/get-started/* ./out/learn
 cp -r ./out/learn/featured-use-cases/* ./out/learn
 cp -r ./out/learn/learn-the-language/* ./out/learn
 cp -r ./out/learn/learn-the-platform/* ./out/learn
+cp -r ./out/policies/* ./out/
 
 #Rearrange files inside the sub-folders of learn-the-platform folder
 cp -r ./out/learn/ballerina-central/* ./out/learn
@@ -14,7 +15,7 @@ cp -r ./out/learn/source-code-dependencies/* ./out/learn
 cp ./out/404/index.html ./out/404.html
 
 #Remove duplicated files
-rm -rf ./out/learn/get-started/ ./out/learn/featured-use-cases/ ./out/learn/learn-the-language/ ./out/learn/learn-the-platform/
+rm -rf ./out/learn/get-started/ ./out/learn/featured-use-cases/ ./out/learn/learn-the-language/ ./out/learn/learn-the-platform/ ./out/policies/
 rm -rf ./out/learn/ballerina-central/ ./out/learn/ballerina-tooling/ ./out/learn/configure-observe/ ./out/learn/java-interoperability/ ./out/learn/test-document-the-code/ ./out/learn/source-code-dependencies/
 
 
@@ -23,6 +24,7 @@ sed -i 's/get-started\///g' ./out/sitemap.xml
 sed -i 's/featured-use-cases\///g' ./out/sitemap.xml
 sed -i 's/learn-the-language\///g' ./out/sitemap.xml
 sed -i 's/learn-the-platform\///g' ./out/sitemap.xml
+sed -i 's/policies\///g' ./out/sitemap.xml
 
 sed -i 's/ballerina-central\///g' ./out/sitemap.xml
 sed -i 's/ballerina-tooling\///g' ./out/sitemap.xml
